Add interface for DataRepository
