+++ composeApp/src/desktopMain/kotlin/des/c5inco/mesh/data/AppDataRepository.kt
@@ -1,7 +1,8 @@
 package des.c5inco.mesh.data
 
-import androidx.compose.ui.graphics.Color
+import data.DataRepository
 import data.MeshState
+import data.defaultPresetColors
 import data.getRoomDatabase
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
@@ -13,17 +14,7 @@ import model.MeshPoint
 import model.SavedColor
 import model.toSavedColor
 
-val defaultPresetColors = listOf(
-    Color(0xff7766EE),
-    Color(0xff8899ff),
-    Color(0xff429BED),
-    Color(0xff4FC1A6),
-    Color(0xffF0C03E),
-    Color(0xffff5599),
-    Color(0xFFFF00FF),
-)
-
-class AppDataRepository {
+class AppDataRepository : DataRepository {
     private val database = getRoomDatabase(getDatabaseBuilder())
     private val scope = CoroutineScope(SupervisorJob() + Dispatchers.IO)
 
@@ -40,43 +31,43 @@ class AppDataRepository {
         }
     }
 
-    fun getAllColors(): Flow<List<SavedColor>> {
+    override fun getAllColors(): Flow<List<SavedColor>> {
         return database.savedColorDao().getAll()
     }
 
-    fun getPresetColors(): Flow<List<SavedColor>> {
+    override fun getPresetColors(): Flow<List<SavedColor>> {
         return database.savedColorDao().getAllPresets()
     }
 
-    fun getCustomColors(): Flow<List<SavedColor>> {
+    override fun getCustomColors(): Flow<List<SavedColor>> {
         return database.savedColorDao().getAllCustom()
     }
 
-    fun addColor(color: SavedColor) {
+    override fun addColor(color: SavedColor) {
         scope.launch {
             database.savedColorDao().insertAll(color)
         }
     }
 
-    fun deleteColor(color: SavedColor) {
+    override fun deleteColor(color: SavedColor) {
         scope.launch {
             database.savedColorDao().delete(color)
         }
     }
 
-    fun loadMeshState(): MeshState {
+    override fun loadMeshState(): MeshState {
         return MeshStateManager.loadState()
     }
 
-    fun saveMeshState(state: MeshState) {
+    override fun saveMeshState(state: MeshState) {
         MeshStateManager.saveState(state)
     }
 
-    fun getMeshPoints(): Flow<List<MeshPoint>> {
+    override fun getMeshPoints(): Flow<List<MeshPoint>> {
         return database.meshPointDao().getAll()
     }
 
-    suspend fun saveMeshPoints(points: List<MeshPoint>) {
+    override suspend fun saveMeshPoints(points: List<MeshPoint>) {
         database.meshPointDao().deleteAll()
         database.meshPointDao().insertAll(*points.toTypedArray())
     }
